requirements.txt
asgiref==3.2.3
cffi==1.14.0
cryptography==2.8
Django==3.0.3
django-rest-knox==4.1.0
djangorestframework==3.11.0
djangorestframework-datatables==0.5.1
mysqlclient==1.4.6
Pillow==7.0.0
pkg-resources==0.0.0
pycparser==2.19
python-decouple==3.3
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
